requirements.txt
pytorch
pandas
librosa
matplotlib
seaborn

